Modificacion de datos de facturacion

+++ app/src/main/java/com/example/elretornodeloretro/model/ProductOrder.kt
@@ -3,6 +3,6 @@ package com.example.elretornodeloretro.model
 data class ProductOrder(
     val NOMBRE_PLATAFORMA: String,
     val NOMBRE_TIPO: String,
-    val PRECIO_FINAL: Int,
+    val PRECIO_FINAL: Float,
     val TITULO: String
 )
